discord_roles/requirements.txt
 mysql-connector-python
 pandas
 numpy
 discord
